pyproject.toml
[tool.poetry]
name = "atlas"
version = "0.1.0"
description = "Atlas Knowledge Base API"
authors = ["User <user@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
pymongo = "^4.6.0"
markitdown = "^0.0.1a2"
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
httpx = "^0.25.2"
python-dotenv = "^1.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]